// src/app/api/model/support-ticket-dto-in.ts
// @ts-nocheck
/**
 * CheckItOut Platform API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TicketCategory } from './ticket-category';


export interface SupportTicketDtoIn { 
    category: TicketCategory;
    contactEmail?: string;
    description?: string;
    subject?: string;
    technicalDescription?: string;
}