// src/app/api/model/ticket-attachment-dto-out.ts
// @ts-nocheck
/**
 * CheckItOut Platform API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TicketAttachmentDtoOut { 
    contentType?: string;
    downloadUrl?: string;
    fileName?: string;
    fileSize?: number;
    id?: number;
    ticketId?: number;
    uploadTime?: string;
}